requirements.txt
pillow
pip